+++ src/app/core/messageType.ts
@@ -1,0 +1,6 @@
+export enum MessageType {
+  BandcampData = 'B2D_BC_DATA',
+  DiscogsEditPageData = 'B2D_DISCOGS_EDIT_PAGE_DATA',
+  Metadata = 'B2D_METADATA',
+  Search = 'B2D_SEARCH'
+}