+++ src/popup/components/console-command.js
@@ -36,13 +36,16 @@ class ConsoleCommand extends HTMLElement {
     self.datalist = self.shadowRoot.querySelector('#command-list');
   }
 
-  addCommand(commandName, handler) {
+  setCommand(commandName, handler) {
     const self = this;
-    self.commands[commandName] = handler;
 
-    const option = document.createElement('option');
-    option.value = commandName;
-    self.datalist.appendChild(option);
+    if (!self.commands[commandName]) {
+      const option = document.createElement('option');
+      option.value = commandName;
+      self.datalist.appendChild(option);
+    }
+
+    self.commands[commandName] = handler;
 
     return self;
   }