+++ src/popup/components/releases-group-list-element.ts
@@ -5,7 +5,7 @@ import {
   hide,
   show
 } from '../../utils/html';
-import { showReleaseCardTab, showReleasesTabContent } from '../modules/main';
+import { showReleaseCard, showReleases } from '../modules/main';
 import { ReleaseItem } from '../../app/releaseItem';
 import { ArtistItem } from '../../app/artistItem';
 import { BandcampItem } from '../../app/bandcampItem';
@@ -70,7 +70,7 @@ export class ReleasesGroupListElement extends HTMLElement {
     if (content instanceof Element) {
       item.appendChild(content);
     } else {
-      item.textContent = content;
+      item.innerHTML = content;
     }
 
     if (onClick !== null) {
@@ -109,7 +109,7 @@ export class ReleasesGroupListElement extends HTMLElement {
     );
 
     if (contentElement) {
-      const url = item.url;
+      const url = item.url.toString();
       self.addItem(
         'artist',
         url,
@@ -134,7 +134,7 @@ export class ReleasesGroupListElement extends HTMLElement {
     );
 
     if (contentElement) {
-      const url = artist.url;
+      const url = artist.url.toString();
       self.addItem(
         'artist',
         url,
@@ -142,7 +142,7 @@ export class ReleasesGroupListElement extends HTMLElement {
         `Show artist/label information`,
         false,
         (event: Event) => {
-          showReleasesTabContent(item, undefined);
+          showReleases(item, undefined);
           event.preventDefault();
         }
       );
@@ -161,7 +161,7 @@ export class ReleasesGroupListElement extends HTMLElement {
     );
 
     if (contentElement) {
-      const url = item.url;
+      const url = item.url.toString();
       self.addItem(
         'release',
         url,
@@ -185,7 +185,7 @@ export class ReleasesGroupListElement extends HTMLElement {
     );
 
     if (contentElement) {
-      const url = release.url;
+      const url = release.url.toString();
       self.addItem(
         'release',
         url,
@@ -193,7 +193,7 @@ export class ReleasesGroupListElement extends HTMLElement {
         `Open release card`,
         false,
         (event: Event) => {
-          showReleaseCardTab(release);
+          showReleaseCard(release);
           event.preventDefault();
         }
       );